Removed useRequest and added PanelTemplate
